sprint11-fix13

+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserService.java
@@ -70,7 +70,7 @@ public class UserService {
         friendStorage.removeFriend(userId, friendId);
     }
 
-    public List<User> getUserFriends(Integer userId) {
+    public List<User> getUserFriends(Long userId) {
         List<Long> idsUsersFromBd = userStorage.getIdUsers();
         if (!(idsUsersFromBd.contains(userId))) {
             throw new ValidationException("В бд нет пользователя с таким id=" + userId);
